make shareable viewmodel
