+++ src/app/src/main/java/com/example/myapplication/ui/facility/FacilityFragment.java
@@ -4,29 +4,35 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.example.myapplication.User;
 import com.example.myapplication.databinding.FacilityScreenFragmentBinding;
 
 public class FacilityFragment extends Fragment {
 
     private FacilityScreenFragmentBinding binding;
+    FacilityViewModel facilityViewModel;
+    User user;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        FacilityViewModel facilityViewModel =
-                new ViewModelProvider(this).get(FacilityViewModel.class);
 
         binding = FacilityScreenFragmentBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textFacility;
-        facilityViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
+    }
+    @Override
+    public void onViewCreated (@NonNull View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        facilityViewModel = new ViewModelProvider(requireActivity()).get(FacilityViewModel.class);
+        user = (User) getArguments().getSerializable("user");
+
+
     }
 
     @Override
