+++ src/app/src/main/java/com/example/myapplication/ui/facility/FacilityViewModel.java
@@ -4,16 +4,14 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
+import com.example.myapplication.User;
+
 public class FacilityViewModel extends ViewModel {
 
-    private final MutableLiveData<String> mText;
+    User organizer;
 
     public FacilityViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+
     }
 
-    public LiveData<String> getText() {
-        return mText;
-    }
 }
